+++ src/CatLib/DevTools/DevKey.cs
@@ -1,0 +1,19 @@
+namespace CatLib.DevTools;
+
+public enum DevKey
+{
+    Up,
+    Down,
+    Left,
+    Right,
+    Enter,
+    Digit1,
+    Digit2,
+    Digit3,
+    Digit4,
+    Digit5,
+    Digit6,
+    Digit7,
+    Digit8,
+    Digit9
+}